change size_t

+++ src/file_op.h
@@ -9,7 +9,7 @@ namespace bible{
 	 * @brief Obtain the length of file.
 	 * @return The length of file.
 	 */
-	int getFileLength(const char* filename);
+	size_t getFileLength(const char* filename);
 
 	///< attention: add \0 to the end of the file.
 	void loadFile(const char * filename, char * &memblock, size_t &filesize);
